pyproject.toml
[tool.poetry]
name = "rxconf"
version = "v0.0.0-alpha.1"
description = "Powerful and flexible library for dynamic configuration management in Python"
authors = ["Andrey Karchevsky <karych@karych.ru>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/realkarych/rxconf/"
repository = "https://github.com/realkarych/rxconf/"
documentation = "https://realkarych.github.io/rxconf/"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["config", "configuration", "settings", "python"]

[tool.poetry.dependencies]
python = ">=3.9"
pyyaml = "^6.0.2"
typing-extensions = "^4.12.2"

[tool.poetry.dev-dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
mypy = "^1.11.1"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pylint = "^3.2.6"
pyright = "^1.1.376"
ruff = "^0.6.1"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.31"
types-PyYAML = "^6.0.12"
coverage-badge = "^1.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
